Add a G2P translator with lexicon lookup and backoff support
The class SequiturGraphemeToPhonemeTranslator holds one or more sequitur models,
indexed by language code, as well as zero or more pronuciation lexica, which are
also indexed by language code.  This allows us to use the new interface method

    GraphemeToPhonemeTranslatorBase.translate(self, text, lang, failure_langs)

to do G2P on text, or words, by first looking up in a fixed lexicon before
performing G2P, and if that fails continue on trying the same for the languages
listed in `failure_langs`.