Add a script for converting the RU MelGAN models to TorchsScript
The script src/scripts/melgan_convert.py can be used to convert the MelGAN
model (.pt) to a stripped down TorchScript model (.pt). By stripped down we mean
a .pt file that contains only the generator, not the discriminator and other
metadata present in the original. The reduction in size is about 90%.