Addition of a local models repository for models that are used in unit testing
The test target now fetches models to a models folder which should be in the root directory. Got rid if the external http_file for sequitur.mdl. For this to work in CI, the CI process must download the required models to the beforementioned models folder in the root directory. This works locally but
for this to work in the GitLab CI pipeline, a script must be executed which creates the models folder and downloads the required models into it. That is fetch_models.sh. It downloads the models from the same source as is done for production, this ensures that we test using the same models as are used in production.
This script, along with CI integration, is currently being worked on. It will also be used to get all of the models in production deployment.

I also set up a test class for LexiconGraphemeToPhonemeTranslator. There is already a single test with more to come.